Refactored and added notifications
Got rid of the .favorites file in favor of calling youtube-dl in a loop. This allows for notifications to be assembled after each successful download.

Notifications are built and issued using pynotify. Thumbnail images are pulled from Youtube links.

Added extra instructions and an example for setting up this script as a cronjob.

Added myself to __author__ list.

Noted this work as a derivative release under the 3-clause BSD license.